Started work on the API, currently the following functions are working: * Addition/Deletion/Fetching of Pages * Addition/Deletion/Fetching of Plugins to a Page * Addition/Deletion/Fetching Plugin Info from a Page
Also started connecting the Frontend to the Backend. There are some caveats, however over-all it looks promising.

Added SailsJS Passport support, but I do not yet know how to use it, so for now it will be installed but not set-up.

Policies are part of the Passport support chunk, that will get some serious work once I understand it's core-features.

Removed one plugin, and kept one so that I can do faster debugging.
module.json will be getting some work to make it easier to quickly add plugins...
Clicking the module name in the pageBuilder will add the module to the page and only that page.  Each page has it's own modules and their own configs, so the same module can be on two different pages and have different settings... More flexibility will be added later.

Blueprints have been disabled because I honestly prefer building my own routes and knowing where all the routes are and what they define. (But they are great for learning... So be sure to try them out! :D)

I have also begun work on the hook aspect of the QuickSale project, so that plugins can hook. However I am trying ot brainstoem easier ways, and currently researching big-open source projects for their methods of implementing plugin-hooks, so it is a WIP.